squelch intel compile warning

+++ src/ref_layer.c
@@ -1072,7 +1072,7 @@ static REF_FCN REF_STATUS ref_layer_prism_insert_hair(REF_GRID ref_grid,
                                       REF_EMPTY, constraining_faceid),
               "ball");
           if (REF_CAVITY_UNKNOWN != ref_cavity_state(ref_cavity)) {
-            printf(" form state %d error\n", ref_cavity_state(ref_cavity));
+            printf(" form state %d error\n", (int)ref_cavity_state(ref_cavity));
             ref_cavity_tec(ref_cavity, "cav-form.tec");
           }
           RSB(ref_cavity_enlarge_combined(ref_cavity), "enlarge", {
@@ -1084,7 +1084,7 @@ static REF_FCN REF_STATUS ref_layer_prism_insert_hair(REF_GRID ref_grid,
                                                     constraining_faceid),
                 "unconst", {
                   printf(" unconstrain state %d\n",
-                         ref_cavity_state(ref_cavity));
+                         (int)ref_cavity_state(ref_cavity));
                   ref_cavity_tec(ref_cavity, "cav-unconst.tec");
                 });
             RSB(ref_cavity_enlarge_combined(ref_cavity), "enlarge", {
